+++ tests/testthat/test-validation-asserts.R
@@ -156,8 +156,27 @@ test_that("can test region codes are consistent", {
 test_that("can check file extensions", {
   expect_true(assert_file_extension("testdata/anc.csv", "csv"))
   expect_true(assert_file_extension("testdata/Botswana2018.PJNZ",
-                                    c("PJNZ", "pjnz", "zip")))
+                                    c("PJNZ", "zip")))
+  expect_true(assert_file_extension("testdata/Botswana2018.pjnz",
+                                    c("PJNZ", "zip")))
   expect_error(assert_file_extension("testdata/anc.csv",
-                                     c("PJNZ", "pjnz", "zip")),
-               "File must be of type PJNZ, pjnz, zip, got type csv.")
+                                     c("PJNZ", "zip")),
+               "File must be of type PJNZ, zip, got type csv.")
+})
+
+test_that("can check region file spectrum codes are valid", {
+  shape <- file_object(file.path("testdata", "malawi.geojson"))
+  json <- hintr_geojson_read(shape)
+  expect_true(assert_region_codes_valid(json))
+
+  mock_contains_property <- mockery::mock(c(FALSE, TRUE, TRUE))
+  with_mock("hintr:::features_contain_property" = mock_contains_property, {
+    expect_true(assert_region_codes_valid(json))
+  })
+
+  mock_contains_property <- mockery::mock(c(FALSE, FALSE, TRUE))
+  with_mock("hintr:::features_contain_property" = mock_contains_property, {
+    expect_error(assert_region_codes_valid(json),
+                 "Shape file contains 2 regions with missing spectrum region code, code can only be missing for country level region.")
+  })
 })